‏מחירון הפיילוט, שחזור סיסמה, אימות מייל והגבלת קצב
חבילות: ‏3/12/30 הזדמנויות ב-19.90 / 49.90 / 99.90 ₪ (scripts/seed_packages.py).
‏price_ils היה שלם, כלומר 19.90 היה נשמר כ-19 — מיגרציה 0010 ל-numeric(10,2).

כפתורי התשלום (אשראי, Bit, ‏PayPal) **אינם מחוברים לסליקה**: הם פותחים חלון
שממנו הלקוח יוצא לוואטסאפ/טלפון/מייל, וההודעה נושאת את החבילה ואת מזהה
החברה כדי שהאדמין ימצא אותה ב-/admin. הטלפון והמייל ב-.env.local — הריפו ציבורי.

שחזור סיסמה ואימות מייל דרך Resend: בלי מפתח המייל נכתב ללוג עם הקישור,
כך שפיתוח ובדיקות אינם שולחים דבר החוצה. המפתח מ-RESEND_API_KEY או
מ-backend/resend.key (ב-.gitignore). האימות אינו חוסם — רק תזכורת במסך.

הגבלת קצב **לפי חשבון** ולא רק לפי IP: נבדק ש-Next אינו מוסיף X-Forwarded-For
ומעביר הלאה את זה שהדפדפן שלח, כלומר ספירה לפי IP הייתה גם עקיפה קלה וגם
דלי אחד לכל הלקוחות. כניסה 10 לחשבון ב-15 דק׳, מיילים 3 לכתובת בשעה,
והרשמה 5 ל-IP בשעה — רק כש-TRUSTED_IP_HEADER מוגדר בפריסה.

נבדק: כל הסוויטה, ‏441 בדיקות (9 חדשות); ובדפדפן — חלון הרכישה, שחזור
סיסמה מקצה לקצה, אימות מייל, וחסימה אחרי 10 ניסיונות דרך Next.

+++ apps/shaked-engine/backend/app/models/package.py
@@ -1,8 +1,9 @@
 import enum
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
-from sqlalchemy import DateTime, Enum, ForeignKey, Index, Integer, String, UniqueConstraint, func
+from sqlalchemy import DateTime, Enum, ForeignKey, Index, Integer, Numeric, String, UniqueConstraint, func
 from sqlalchemy.dialects.postgresql import JSONB, UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -17,7 +18,8 @@ class Package(Base):
     id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     name: Mapped[str] = mapped_column(String(255), nullable=False)
     credits: Mapped[int] = mapped_column(Integer, nullable=False)
-    price_ils: Mapped[int] = mapped_column(Integer, nullable=False)
+    # שקלים ואגורות: 19.90, לא 19 ולא 1990.
+    price_ils: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     created_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), server_default=func.now())
